skills/asset-gen/requirements.txt
# Asset Generation Skill Dependencies
google-genai>=0.8.0
Pillow>=10.0.0,<12.0.0
